pyproject.toml
[project]
name = "job-application-tool"
version = "0.1.0"
description = "Automated LinkedIn job application tool with Apify scraping, LLM ranking, and CV generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic-ai>=0.1.0",

    # Database
    "motor>=3.7.0",
    "pymongo>=4.0.0",

    # HTTP Client (for Apify API)
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # LLM Providers
    "openai>=1.0.0",

    # PDF Generation
    "reportlab>=4.0.0",
    "weasyprint>=62.0",

    # CLI & Logging
    "click>=8.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # Async utilities
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
